增加model layout  templete category之间的关联关系，完善发布流程和机制，完善创传图片尺寸问题，新增更改templete时候，category和content自动更改，后续在完善统一发布机制

+++ app/Http/Controllers/Admin/Traits/TagsManage.php
@@ -49,7 +49,8 @@ trait TagsManage
         $exits_tags = $tag_builder->get()->flatMap(function($item){
             return [$item->tags->name];
         })->toArray();
-        if(array_diff($exits_tags, $request->get('tags'))){
+
+        if(!empty($request->get('tags')) && array_diff($exits_tags, $request->get('tags'))){
             $tag_builder->forceDelete();
             $this->saveTags($request, $data, $model_id);
         }
